+++ CHANGES.rst
@@ -24,6 +24,13 @@ New Features
     now possible to specify the table alignment within the text via the
     ``tablealign`` keyword. [#1838]
 
+  - If ``header_start`` is specified in a call to ``ascii.get_reader`` or any 
+    method that calls ``get_reader`` (e.g. ``ascii.read``) but ``data_start``
+    is not specified at the same time, then ``data_start`` is calculated so
+    that the data starts after the header. Before this, the default was 
+    that the header line was read again as the first data line 
+    [#855 and #1844].
+
 - ``astropy.io.fits``
 
 - ``astropy.io.misc``